Созданы файлы класса ntdisplay для работы с дисплеем
